auth

+++ app/Http/Controllers/FirmenController.php
@@ -4,12 +4,22 @@ namespace App\Http\Controllers;
 
 use App\firmen;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class FirmenController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     public static function asArray()
     {
-        return firmen::all();
+        if (in_array(Auth::user()->typ, ['admin', 'employe'])) {
+            return firmen::all();
+        } else {
+            return view('welcome');
+        }
     }
 
     /**
@@ -19,7 +29,11 @@ class FirmenController extends Controller
      */
     public function create()
     {
-        return view('firmen.create');
+        if (in_array(Auth::user()->typ, ['admin', 'employe'])) {
+            return view('firmen.create');
+        } else {
+            return view('welcome');
+        }
     }
 
     /**
@@ -30,29 +44,33 @@ class FirmenController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'name' => 'required',
-            'ort' => 'required',
-            'plz' => 'bail|numeric|between:0,99999',
-            'strasse' => 'required'
-        ]);
+        if (in_array(Auth::user()->typ, ['admin', 'employe'])) {
+            $request->validate([
+                'name' => 'required',
+                'ort' => 'required',
+                'plz' => 'bail|numeric|between:0,99999',
+                'strasse' => 'required'
+            ]);
 
-        $firma = new firmen;
-        $firma->Firmenname = request('name');
-        $firma->Firmenbezeichnung = request('bezeichnung');
-        $firma->Firmenwebseite = request('webseite');
-        $firma->Email = request('email');
-        $firma->Ort = request('ort');
-        $firma->PLZ = request('plz');
-        $firma->Strasse = request('strasse');
-        $firma->Telefon = request('telefon');
-        try {
-            $firma->save();
-        } catch (\Exception $e) {
-            return view('firmen.create')->withErrors($e->getMessage());
+            $firma = new firmen;
+            $firma->Firmenname = request('name');
+            $firma->Firmenbezeichnung = request('bezeichnung');
+            $firma->Firmenwebseite = request('webseite');
+            $firma->Email = request('email');
+            $firma->Ort = request('ort');
+            $firma->PLZ = request('plz');
+            $firma->Strasse = request('strasse');
+            $firma->Telefon = request('telefon');
+            try {
+                $firma->save();
+            } catch (\Exception $e) {
+                return view('firmen.create')->withErrors($e->getMessage());
+            }
+
+            return redirect(route('firmen.show', $firma));
+        } else {
+            return view('welcome');
         }
-
-        return redirect(route('firmen.show', $firma));
     }
 
     /**
@@ -63,70 +81,86 @@ class FirmenController extends Controller
      */
     public function show(firmen $firmen)
     {
-        return view('firmen.show', compact('firmen'));
+        if (in_array(Auth::user()->typ, ['admin', 'employe'])) {
+            return view('firmen.show', compact('firmen'));
+        } else {
+            return view('welcome');
+        }
     }
 
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\firmen  $firmen
+     * @param  \App\firmen $firmen
      * @return \Illuminate\Http\Response
      */
     public function edit(firmen $firmen)
     {
-        return view('firmen.edit', compact('firmen'));
+        if (in_array(Auth::user()->typ, ['admin', 'employe'])) {
+            return view('firmen.edit', compact('firmen'));
+        } else {
+            return view('welcome');
+        }
     }
 
     /**
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request $request
-     * @param  \App\firmen  $firmen
+     * @param  \App\firmen $firmen
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, firmen $firmen)
     {
-        $request->validate([
-            'name' => 'required',
-            'ort' => 'required',
-            'plz' => 'bail|numeric|between:0,99999',
-            'strasse' => 'required'
-        ]);
+        if (in_array(Auth::user()->typ, ['admin', 'employe'])) {
+            $request->validate([
+                'name' => 'required',
+                'ort' => 'required',
+                'plz' => 'bail|numeric|between:0,99999',
+                'strasse' => 'required'
+            ]);
 
-        $firmen->update(array(
-            'Firmenname' => request('name'),
-            'Firmenbezeichnung' => request('bezeichnung'),
-            'Firmenwebseite' => request('webseite'),
-            'Email' => request('email'),
-            'Ort' => request('ort'),
-            'PLZ' => request('plz'),
-            'Strasse' => request('strasse'),
-            'Telefon' => request('telefon')
-        ));
-        try {
-            $firmen->save();
-        } catch (\Exception $e) {
-            return view('firmen.edit', $firmen)->withErrors($e->getMessage());
+            $firmen->update(array(
+                'Firmenname' => request('name'),
+                'Firmenbezeichnung' => request('bezeichnung'),
+                'Firmenwebseite' => request('webseite'),
+                'Email' => request('email'),
+                'Ort' => request('ort'),
+                'PLZ' => request('plz'),
+                'Strasse' => request('strasse'),
+                'Telefon' => request('telefon')
+            ));
+            try {
+                $firmen->save();
+            } catch (\Exception $e) {
+                return view('firmen.edit', $firmen)->withErrors($e->getMessage());
+            }
+
+            return redirect(route('firmen.show', $firmen));
+        } else {
+            return view('welcome');
         }
-
-        return redirect(route('firmen.show', $firmen));
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\firmen  $firmen
+     * @param  \App\firmen $firmen
      * @return \Illuminate\Http\Response
      */
     public function destroy(firmen $firmen)
     {
-        try {
-            $firmen->delete();
-        } catch (\Exception $e) {
-            return view('firmen.show', compact('firmen'))->withErrors($e->getMessage());
-        }
+        if (in_array(Auth::user()->typ, ['admin', 'employe'])) {
+            try {
+                $firmen->delete();
+            } catch (\Exception $e) {
+                return view('firmen.show', compact('firmen'))->withErrors($e->getMessage());
+            }
 
-        return $this->index();
+            return $this->index();
+        } else {
+            return view('welcome');
+        }
     }
 
     /**
@@ -136,7 +170,11 @@ class FirmenController extends Controller
      */
     public function index()
     {
-        $firmen = firmen::paginate(25);
-        return view('firmen.index', compact('firmen'));
+        if (in_array(Auth::user()->typ, ['admin', 'employe'])) {
+            $firmen = firmen::paginate(25);
+            return view('firmen.index', compact('firmen'));
+        } else {
+            return view('welcome');
+        }
     }
 }
