+++ app/Http/Controllers/PraktikaController.php
@@ -2,10 +2,9 @@
 
 namespace App\Http\Controllers;
 
-use App\ansprechpartnerliste;
 use App\praktika;
 use Illuminate\Http\Request;
-use TheSeer\Tokenizer\Exception;
+use Illuminate\Support\Facades\Auth;
 
 class Praktikacontroller extends Controller
 {
@@ -22,7 +21,11 @@ class Praktikacontroller extends Controller
      */
     public function create()
     {
-        return view('praktika.create');
+        if (in_array(Auth::user()->typ, ['admin', 'employe'])) {
+            return view('praktika.create');
+        } else {
+            return view('welcome');
+        }
     }
 
     /**
@@ -33,24 +36,28 @@ class Praktikacontroller extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'teilnehmer' => 'exists:teilnehmer,Teilnehmer_ID',
-            'firma' => 'exists:firmen,Firmen_ID',
-            'zeit' => 'exists:praktikazeitraeume,Praktikumszeit_ID',
-            'status' => 'required'
-        ]);
+        if (in_array(Auth::user()->typ, ['admin', 'employe'])) {
+            $request->validate([
+                'teilnehmer' => 'exists:teilnehmer,Teilnehmer_ID',
+                'firma' => 'exists:firmen,Firmen_ID',
+                'zeit' => 'exists:praktikazeitraeume,Praktikumszeit_ID',
+                'status' => 'required'
+            ]);
 
-        $praktikum = new praktika;
-        $praktikum->Teilnehmer_ID = request('teilnehmer');
-        $praktikum->Firmen_ID = request('firma');
-        $praktikum->Praktikumszeit_ID = request('zeit');
-        $praktikum->Status = request('status');
-        try {
-            $praktikum->save();
-        } catch (\Exception $e) {
-            return view('praktika.create')->withErrors($e->getMessage());
+            $praktikum = new praktika;
+            $praktikum->Teilnehmer_ID = request('teilnehmer');
+            $praktikum->Firmen_ID = request('firma');
+            $praktikum->Praktikumszeit_ID = request('zeit');
+            $praktikum->Status = request('status');
+            try {
+                $praktikum->save();
+            } catch (\Exception $e) {
+                return view('praktika.create')->withErrors($e->getMessage());
+            }
+            return redirect(route('praktika.show', $praktikum));
+        } else {
+            return view('welcome');
         }
-        return redirect(route('praktika.show', $praktikum));
     }
 
     /**
@@ -61,67 +68,82 @@ class Praktikacontroller extends Controller
      */
     public function show(praktika $praktika)
     {
-        return view('praktika.show', compact('praktika'));
+        if (in_array(Auth::user()->typ, ['admin', 'employe'])) {
+            return view('praktika.show', compact('praktika'));
+        } else {
+            return view('welcome');
+        }
     }
 
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\praktika  $praktika
+     * @param  \App\praktika $praktika
      * @return \Illuminate\Http\Response
      */
     public function edit(praktika $praktika)
     {
+        if (in_array(Auth::user()->typ, ['admin', 'employe'])) {
 
-        return view('praktika.edit', compact('praktika'));
+            return view('praktika.edit', compact('praktika'));
+        } else {
+            return view('welcome');
+        }
     }
 
     /**
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request $request
-     * @param  \App\praktika  $praktika
+     * @param  \App\praktika $praktika
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, praktika $praktika)
     {
-        $request->validate([
-            'teilnehmer' => 'exists:teilnehmer,Teilnehmer_ID',
-            'firma' => 'exists:firmen,Firmen_ID',
-            'zeit' => 'exists:praktikazeitraeume,Praktikumszeit_ID',
-            'status' => 'required'
-        ]);
+        if (in_array(Auth::user()->typ, ['admin', 'employe'])) {
+            $request->validate([
+                'teilnehmer' => 'exists:teilnehmer,Teilnehmer_ID',
+                'firma' => 'exists:firmen,Firmen_ID',
+                'zeit' => 'exists:praktikazeitraeume,Praktikumszeit_ID',
+                'status' => 'required'
+            ]);
 
-        $praktika->update(array(
-            'Teilnehmer_ID' => request('teilnehmer'),
-            'Firmen_ID' => request('firma'),
-            'Praktikumszeit_ID' => request('zeit'),
-            'Status' => request('status')
-        ));
-        try {
-            $praktika->save();
-        } catch (\Exception $e) {
-            return view('praktika.edit', $praktika)->withErrors($e->getMessage());
+            $praktika->update(array(
+                'Teilnehmer_ID' => request('teilnehmer'),
+                'Firmen_ID' => request('firma'),
+                'Praktikumszeit_ID' => request('zeit'),
+                'Status' => request('status')
+            ));
+            try {
+                $praktika->save();
+            } catch (\Exception $e) {
+                return view('praktika.edit', $praktika)->withErrors($e->getMessage());
+            }
+
+            return redirect(route('praktika.show', $praktika));
+        } else {
+            return view('welcome');
         }
-
-        return redirect(route('praktika.show', $praktika));
-
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\praktika  $praktika
+     * @param  \App\praktika $praktika
      * @return \Illuminate\Http\Response
      */
     public function destroy(praktika $praktika)
     {
-        try {
-            $praktika->delete();
-        } catch (\Exception $e) {
-            return view('praktika.show', compact('praktika'))->withErrors($e->getMessage());
+        if (in_array(Auth::user()->typ, ['admin', 'employe'])) {
+            try {
+                $praktika->delete();
+            } catch (\Exception $e) {
+                return view('praktika.show', compact('praktika'))->withErrors($e->getMessage());
+            }
+            return $this->index();
+        } else {
+            return view('welcome');
         }
-        return $this->index();
     }
 
     /**
@@ -131,7 +153,11 @@ class Praktikacontroller extends Controller
      */
     public function index()
     {
-        $praktika = praktika::paginate(25);
-        return view('praktika.index', compact('praktika'));
+        if (in_array(Auth::user()->typ, ['admin', 'employe'])) {
+            $praktika = praktika::paginate(25);
+            return view('praktika.index', compact('praktika'));
+        } else {
+            return view('welcome');
+        }
     }
 }
